Move each composable to respective view package  * Also configures error card to adaptable theming
