+++ app/src/main/java/com/farmingdale/stockscreener/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.farmingdale.stockscreener.ui.theme.StockScreenerTheme
-import com.farmingdale.stockscreener.views.MainView
+import com.farmingdale.stockscreener.views.main.MainView
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
