+++ app/src/main/java/com/farmingdale/stockscreener/views/home/ErrorCard.kt
@@ -1,4 +1,4 @@
-package com.farmingdale.stockscreener.views
+package com.farmingdale.stockscreener.views.home
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -8,17 +8,16 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Warning
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
@@ -39,7 +38,7 @@ fun ErrorCard(
                 .clickable { refresh() },
             shape = RoundedCornerShape(10),
             colors = CardDefaults.cardColors(
-                containerColor = Color(0xFFFFE4E1)
+                containerColor = MaterialTheme.colorScheme.errorContainer
             )
         ) {
             Column(
@@ -54,8 +53,9 @@ fun ErrorCard(
                 Text(
                     text = stringResource(id = R.string.notLoaded),
                     modifier = Modifier.padding(16.dp),
-                    style = MaterialTheme.typography.body2,
-                    textAlign = TextAlign.Center
+                    style = MaterialTheme.typography.bodyMedium,
+                    textAlign = TextAlign.Center,
+                    color = MaterialTheme.colorScheme.onErrorContainer
                 )
             }
         }
