+++ app/src/main/java/com/farmingdale/stockscreener/views/home/MarketIndices.kt
@@ -1,4 +1,4 @@
-package com.farmingdale.stockscreener.views
+package com.farmingdale.stockscreener.views.home
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
